feat: refactor Chart component to accept props for data and configuration
