+++ paystell-frontend/src/components/dashboard/sales/SalesHistory.tsx
@@ -9,7 +9,29 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
+import { ChartConfig } from "@/components/ui/chart";
 const SalesHistory = () => {
+  const chartData = [
+    { month: "January", desktop: 186, mobile: 80 },
+    { month: "February", desktop: 305, mobile: 200 },
+    { month: "March", desktop: 237, mobile: 120 },
+    { month: "April", desktop: 73, mobile: 190 },
+    { month: "May", desktop: 209, mobile: 130 },
+    { month: "June", desktop: 214, mobile: 140 },
+    { month: "July", desktop: 214, mobile: 140 },
+  ];
+
+  const chartConfig = {
+    desktop: {
+      label: "Snacks",
+      color: "#2563eb",
+    },
+    mobile: {
+      label: "drinks",
+      color: "#60a5fa",
+    },
+  } satisfies ChartConfig;
+
   return (
     <div className="mt-6 flex flex-col  md:flex-row">
       <Table className="md:w-[90%]">
@@ -71,7 +93,7 @@ const SalesHistory = () => {
         </TableBody>
       </Table>
       <div className="mt-5 md:w-[40%]">
-        <Chart />
+        <Chart chartData={chartData} chartConfig={chartConfig} />
       </div>
     </div>
   );
